requirements.txt
asgiref==3.8.1
boto3==1.35.0
botocore==1.35.0
contourpy==1.2.1
cycler==0.12.1
Django==5.1
djangorestframework==3.15.2
drf-yasg==1.21.7
fonttools==4.53.1
inflection==0.5.1
jmespath==1.0.1
kiwisolver==1.4.5
matplotlib==3.9.2
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
psycopg2==2.9.9
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==0.19.1
pytz==2024.1
PyYAML==6.0.2
s3transfer==0.10.2
setuptools==72.1.0
six==1.16.0
SQLAlchemy==2.0.32
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
